Extraindo testes genéricos para classe de teste de BaseGraph
